REF: Remove `_BaseTest` construct
