+++ .pre-commit-config.yaml
@@ -23,6 +23,10 @@ repos:
   hooks:
   - id: isort
     files: ^openff|plugins|stubs
+- repo: https://github.com/asottile/yesqa
+  rev: v1.5.0
+  hooks:
+  - id: yesqa
 - repo: https://github.com/PyCQA/flake8
   rev: 7.0.0
   hooks:
