+++ openff/interchange/_tests/conftest.py
@@ -1,9 +1,15 @@
 """Pytest configuration."""
 
+import numpy
 import pytest
 from openff.toolkit import ForceField, Molecule, Topology
-from openff.toolkit.typing.engines.smirnoff.parameters import BondType, VirtualSiteType
+from openff.toolkit.typing.engines.smirnoff.parameters import (
+    BondType,
+    ChargeIncrementModelHandler,
+    VirtualSiteType,
+)
 from openff.units import Quantity, unit
+from openff.utilities import get_data_file_path
 
 from openff.interchange._tests import MoleculeWithConformer, get_test_file_path
 
@@ -370,3 +376,166 @@ def cyclohexane() -> Molecule:
     cyclohexane.add_bond(5, 17, 1, False)
 
     return cyclohexane
+
+
+@pytest.fixture(autouse=True)
+def _initdir(tmpdir):
+    tmpdir.chdir()
+
+
+@pytest.fixture()
+def ethanol_top(ethanol):
+    """Fixture that builds a simple four ethanol topology."""
+    return Topology.from_molecules(4 * [ethanol])
+
+
+@pytest.fixture()
+def mainchain_ala():
+    molecule = Molecule.from_file(
+        get_data_file_path("proteins/MainChain_ALA.sdf", "openff.toolkit"),
+    )
+    molecule._add_default_hierarchy_schemes()
+    molecule.perceive_residues()
+    molecule.perceive_hierarchy()
+
+    return molecule
+
+
+@pytest.fixture()
+def mainchain_arg():
+    molecule = Molecule.from_file(
+        get_data_file_path("proteins/MainChain_ARG.sdf", "openff.toolkit"),
+    )
+    molecule._add_default_hierarchy_schemes()
+    molecule.perceive_residues()
+    molecule.perceive_hierarchy()
+
+    return molecule
+
+
+@pytest.fixture()
+def two_peptides(mainchain_ala, mainchain_arg):
+    return Topology.from_molecules([mainchain_ala, mainchain_arg])
+
+
+@pytest.fixture()
+def xml_ff_bo_bonds() -> str:
+    return """<?xml version='1.0' encoding='ASCII'?>
+<SMIRNOFF version="0.3" aromaticity_model="OEAroModel_MDL">
+  <Bonds version="0.3" fractional_bondorder_method="AM1-Wiberg" fractional_bondorder_interpolation="linear">
+    <Bond smirks="[#6:1]~[#8:2]" id="bbo1"
+        k_bondorder1="100.0*kilocalories_per_mole/angstrom**2"
+        k_bondorder2="1000.0*kilocalories_per_mole/angstrom**2"
+        length_bondorder1="1.5*angstrom"
+        length_bondorder2="1.0*angstrom"/>
+  </Bonds>
+</SMIRNOFF>
+"""
+
+
+@pytest.fixture()
+def xml_ff_bo() -> str:
+    return """<?xml version='1.0' encoding='ASCII'?>
+<SMIRNOFF version="0.3" aromaticity_model="OEAroModel_MDL">
+  <Bonds version="0.3" fractional_bondorder_method="AM1-Wiberg"
+    fractional_bondorder_interpolation="linear">
+    <Bond
+      smirks="[#6X4:1]~[#8X2:2]"
+      id="bbo1"
+      k_bondorder1="101.0 * kilocalories_per_mole/angstrom**2"
+      k_bondorder2="123.0 * kilocalories_per_mole/angstrom**2"
+      length_bondorder1="1.4 * angstrom"
+      length_bondorder2="1.3 * angstrom"
+      />
+  </Bonds>
+  <ProperTorsions version="0.3" potential="k*(1+cos(periodicity*theta-phase))">
+    <Proper smirks="[*:1]~[#6X3:2]~[#6X3:3]~[*:4]" id="tbo1" periodicity1="2" phase1="0.0 * degree"
+    k1_bondorder1="1.00*kilocalories_per_mole" k1_bondorder2="1.80*kilocalories_per_mole" idivf1="1.0"/>
+    <Proper smirks="[*:1]~[#6X4:2]~[#8X2:3]~[*:4]" id="tbo2" periodicity1="2" phase1="0.0 * degree"
+    k1_bondorder1="1.00*kilocalories_per_mole" k1_bondorder2="1.80*kilocalories_per_mole" idivf1="1.0"/>
+  </ProperTorsions>
+</SMIRNOFF>
+"""
+
+
+@pytest.fixture()
+def methane():
+    return Molecule.from_smiles("C")
+
+
+@pytest.fixture()
+def parsley():
+    return ForceField("openff-1.0.0.offxml")
+
+
+@pytest.fixture()
+def hydrogen_cyanide():
+    return Molecule.from_mapped_smiles("[H:1][C:2]#[N:3]")
+
+
+@pytest.fixture()
+def hydrogen_cyanide_reversed():
+    return Molecule.from_mapped_smiles("[H:3][C:2]#[N:1]")
+
+
+@pytest.fixture()
+def hexane_diol():
+    molecule = Molecule.from_smiles("OCCCCCCO")
+    molecule.assign_partial_charges(partial_charge_method="gasteiger")
+    molecule.partial_charges.m
+    return molecule
+
+
+@pytest.fixture()
+def hydrogen_chloride():
+    return Molecule.from_mapped_smiles("[Cl:1][H:2]")
+
+
+@pytest.fixture()
+def formaldehyde():
+    return Molecule.from_mapped_smiles("[H:3][C:1]([H:4])=[O:2]")
+
+
+@pytest.fixture()
+def acetaldehyde():
+    return Molecule.from_mapped_smiles(
+        "[C:1]([C:2](=[O:3])[H:7])([H:4])([H:5])[H:6]",
+    )
+
+
+@pytest.fixture()
+def methane_with_conformer(methane):
+    methane.add_conformer(
+        unit.Quantity(
+            numpy.random.random((methane.n_atoms, 3)),
+            unit.angstrom,
+        ),
+    )
+    return methane
+
+
+@pytest.fixture()
+def ethanol_with_conformer(ethanol):
+    ethanol.add_conformer(
+        unit.Quantity(
+            numpy.random.random((ethanol.n_atoms, 3)),
+            unit.angstrom,
+        ),
+    )
+    return ethanol
+
+
+@pytest.fixture()
+def cb8_host() -> Molecule:
+    return Molecule.from_file(get_test_file_path("CB8.sdf"))
+
+
+@pytest.fixture()
+def no_charges() -> ForceField:
+    sage = ForceField("openff_unconstrained-2.0.0.offxml")
+    sage.deregister_parameter_handler("ToolkitAM1BCC")
+    sage.register_parameter_handler(
+        ChargeIncrementModelHandler(version=0.3, partial_charge_method="formal_charge"),
+    )
+
+    return sage
