+++ openff/interchange/_tests/interoperability_tests/internal/test_amber.py
@@ -11,7 +11,6 @@ from openff.utilities import (
 )
 
 from openff.interchange import Interchange
-from openff.interchange._tests import _BaseTest
 from openff.interchange.drivers import get_amber_energies, get_openmm_energies
 
 if has_package("openmm"):
@@ -20,7 +19,7 @@ if has_package("openmm"):
     import openmm.unit
 
 
-class TestAmber(_BaseTest):
+class TestAmber:
     @pytest.mark.skip(reason="Need replacement route to reference positions")
     def test_inpcrd(self, sage):
         mol = Molecule.from_smiles(10 * "C")
@@ -86,7 +85,7 @@ class TestAmber(_BaseTest):
         omm_energies.compare(amb_energies)
 
 
-class TestPRMTOP(_BaseTest):
+class TestPRMTOP:
     @skip_if_missing("mdtraj")
     @skip_if_missing("MDAnalysis")
     @skip_if_missing("openmm")
@@ -124,7 +123,7 @@ class TestPRMTOP(_BaseTest):
         assert mdtraj_atom_names == pdb_atom_names
 
 
-class TestAmberResidues(_BaseTest):
+class TestAmberResidues:
     @pytest.mark.parametrize("patch_residue_name", [True, False])
     def test_single_residue_system_residue_name(
         self,
