+++ openff/interchange/_tests/unit_tests/common/test_positions.py
@@ -1,33 +1,11 @@
 import numpy
-import pytest
 from openff.toolkit import Topology
 from openff.units import unit
 
-from openff.interchange._tests import _BaseTest
 from openff.interchange.common._positions import _infer_positions
 
 
-class TestInferPositions(_BaseTest):
-    @pytest.fixture()
-    def methane_with_conformer(self, methane):
-        methane.add_conformer(
-            unit.Quantity(
-                numpy.random.random((methane.n_atoms, 3)),
-                unit.angstrom,
-            ),
-        )
-        return methane
-
-    @pytest.fixture()
-    def ethanol_with_conformer(self, ethanol):
-        ethanol.add_conformer(
-            unit.Quantity(
-                numpy.random.random((ethanol.n_atoms, 3)),
-                unit.angstrom,
-            ),
-        )
-        return ethanol
-
+class TestInferPositions:
     def test_short_circuit(self, methane):
         positions = unit.Quantity(
             numpy.random.random((methane.n_atoms, 3)),
