+++ openff/interchange/_tests/unit_tests/drivers/test_report.py
@@ -3,7 +3,6 @@ from openff.units import unit
 from openff.units.openmm import ensure_quantity
 from openff.utilities.testing import skip_if_missing
 
-from openff.interchange._tests import _BaseTest
 from openff.interchange.constants import kj_mol
 from openff.interchange.drivers.report import EnergyReport
 from openff.interchange.exceptions import (
@@ -13,7 +12,7 @@ from openff.interchange.exceptions import (
 )
 
 
-class TestEnergyReport(_BaseTest):
+class TestEnergyReport:
     @pytest.fixture()
     def report(self):
         return EnergyReport(
